+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unpublish
 
+### Features
+
+- Add function for deleting deleted tasks, #16
+
 ## 0.0.1
+
+### Features
 
 - Initialize Project